requirements.txt
amqp==5.3.1
billiard==4.2.4
blinker==1.9.0
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.6.2
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.12.1
pytest==9.0.2
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.3.0
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
Werkzeug==3.1.6
